[Feat] 월별 수익률 조회 기능 추가 (#77)
* [Refactor] 변동성 데이터 타입 변경

* [Feat] 월별 수익률 관련 로직 작성

+++ src/main/java/kwangwoon/chambit/dontworry/domain/stock/domain/Stock.java
@@ -6,6 +6,8 @@ import lombok.*;
 import org.hibernate.annotations.OnDelete;
 import org.hibernate.annotations.OnDeleteAction;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -17,11 +19,11 @@ public class Stock {
 
     private String stockCode;
     private String stockName;
-    private Long volatility;
+    private BigDecimal volatility;
     private String imageUrl;
 
     @Builder
-    public Stock(String stockCode, String stockName, Long volatility, String imageUrl){
+    public Stock(String stockCode, String stockName, BigDecimal volatility, String imageUrl){
         this.stockCode = stockCode;
         this.stockName = stockName;
         this.volatility = volatility;
